% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusteLevenberg.R
\name{ajusteLevenberg}
\alias{ajusteLevenberg}
\title{Ajuste dos mínimos quadrados de Levenberg-Marquardt.}
\usage{
ajusteLevenberg(
  c_infinito,
  k,
  tzero,
  dados_curva,
  real_cont_fw,
  medida,
  tempoB,
  mainNameB,
  sexo = NULL,
  genero = NULL
)
}
\arguments{
\item{c_infinito}{Contém o comprimento assintótico.}

\item{k}{Contém a taxa de crescimento.}

\item{tzero}{Contém o tempo no comprimento zero.}

\item{dados_curva}{Contém o dataframe com os dados das curvas de crescimento.}

\item{real_cont_fw}{Contém do "n" da amostra.}

\item{medida}{Legenda do gráfico eixo y.}

\item{tempoB}{Legenda do gráfico eixo x.}

\item{mainNameB}{Legenda do gráfico título.}

\item{sexo}{Contém um texto referente ao grupo atual.}

\item{genero}{Contém a legenda do grupo atual.}
}
\description{
Script que ajusta as curvas de crescimento de acordo com o modelo de Levenberg-Marquardt.
}
\examples{
ajusteLevenberg(c_infinito, k, tzero, dados_curva, real_cont_fw, medida, tempoB, mainNameB,sexo, genero)

}
