% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvasCrescimento.R
\name{crescimento}
\alias{crescimento}
\title{Curvas de Crescimento.}
\usage{
crescimento(
  dados,
  c_inf,
  B,
  tzer,
  contador,
  tempoB,
  medida,
  mainNameA,
  sexo = NULL
)
}
\arguments{
\item{dados}{Contém os dados para a curva obtidos através do modelo de Ford-Walford.}

\item{c_inf}{Contém o comprimento assintótico.}

\item{B}{Contém a taxa de crescimento.}

\item{tzer}{Contém o tempo no comprimento zero.}

\item{contador}{Contém o "n" da amostra.}

\item{tempoB}{Legenda do gráfico eixo x.}

\item{medida}{Legenda do gráfico eixo y.}

\item{mainNameA}{Legenda do gráfico título.}

\item{sexo}{Contém um texto referente ao grupo atual.}
}
\description{
Script que calcula a curva de crescimento a partir dos parâmetros gerados pelo modelo de Ford-Walford.
}
\examples{
crescimento(dados, c_inf, B, tzer, contador, tempoB, medida, mainNameA, sexo)

}
