% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{cvError_comput}
\alias{cvError_comput}
\title{Compute CV errors}
\usage{
cvError_comput(splitedData, dataTypes, alphas, ThetaHat, d)
}
\arguments{
\item{splitedData}{output of function \code{dataSplit}}

\item{dataTypes}{the data types for each data set}

\item{alphas}{dispersion parameters for each data set}

\item{ThetaHat}{estimated Theta}

\item{d}{a numeric vector contains the number of variables of data sets}
}
\value{
This function returns a vector contains CV errors
}
\description{
This function will compute CV errors for a specific model
}
\examples{
\dontrun{cvError_comput(splitedData,dataTypes,alphas,ThetaHat,d)}
}
