% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_simu_element_sparse.R
\name{dataSimu_element_sparse}
\alias{dataSimu_element_sparse}
\title{Simulate multiple data sets with element-wise sparse patterns}
\usage{
dataSimu_element_sparse(n, ds, R = 3, dataTypes = "GGG",
  noises = rep(1, 3), margProb = 0.1, sparse_ratio = 0.5,
  SNRs = rep(1, 3))
}
\arguments{
\item{n}{the number of objects}

\item{ds}{a vector for the number of variables in each data set}

\item{R}{the number of simulated PCs}

\item{dataTypes}{a string indicates the data type of each data set,
possible options include 'G': Gaussian, 'B': Bernoulli.}

\item{noises}{noise levels of simulated data sets}

\item{margProb}{desired marginal probability for binary data simulation,
used to simulate imbalanced binary data.}

\item{sparse_ratio}{controls the sparse level of element-wise sparsity}

\item{SNRs}{the SNRs for the simulation of the multiple data sets}
}
\value{
Refer the return of thefunction \code{dataSimu_group_sparse}.
}
\description{
This function can generate multiple data sets of mixed data types.
Details of the simultion process can be found in the paper
\url{https://arxiv.org/abs/1902.06241}. In this function, only the
element-wise sparsity is included into the loading matrix.
}
\examples{
\dontrun{
dataSimulation <- dataSimu_element_sparse(n,ds,R,
                                   dataTypes='GGG',
                                   noises=rep(1,3),
                                   margProb=0.1,sparse_ratio=0.5,
                                   SNRs=rep(1,3))
}

}
