% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_simu_group_sparse.R
\name{dataSimu_group_sparse}
\alias{dataSimu_group_sparse}
\title{Simulate multiple data sets with group and element-wise sparse patterns}
\usage{
dataSimu_group_sparse(n, ds, dataTypes = "GGG", noises = rep(1, 3),
  margProb = 0.1, sparse_ratio = 0, SNRgc = 1, SNRlc = rep(1, 3),
  SNRd = rep(1, 3))
}
\arguments{
\item{n}{the number of objects}

\item{ds}{a vector for the number of variables in each data set}

\item{dataTypes}{a string indicates the data type of each data set,
possible options include 'G': Gaussian, 'B': Bernoulli.}

\item{noises}{noise levels of simulated data sets}

\item{margProb}{desired marginal probability for binary data simulation,
used to simulate imbalanced binary data.}

\item{sparse_ratio}{controls the sparse level of element-wise sparsity}

\item{SNRgc}{SNR for global common structure}

\item{SNRlc}{SNRs of the local common structures}

\item{SNRd}{SNRs of the distinct structures}
}
\value{
A list contains the simulated data sets and the simulated parameters. \itemize{
\item X: a list contains the simulated multiple data sets;
\item Theta_simu: simulated natural parameter matrix Theta;
\item mu_simu: simulated offset term mu;
\item U_simu: simulated U;
\item D_simu: simulated D;
\item V_simu: simulated V;
\item E_simu: simulated noise term E;
\item dataTypes: vector form simulated data types;
\item S_simu: desired group sparse pattern;
\item SNRs: used SNRs in simulating common (global, local) and distinct structures;
\item varExpTotals_simu: variation explained ratios for each data set computed
using the simulated parameters;
\item varExpPCs_simu: variation explained raitos for each PC for each data set
computed using the simulated parameters;
}
}
\description{
This function can generate multiple data sets of mixed data types.
Details of the simultion process can be found in the paper
\url{https://arxiv.org/abs/1902.06241}. In the current version, both
group and element-wise sparsity are included into the loading matrix.
}
\examples{
\dontrun{
dataSimulation <- dataSimu_group_sparse(n=200,
                                       ds=c(400,200,100),
                                       dataTypes='GGB')
}

}
