% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{eval_metrics_simu_group}
\alias{eval_metrics_simu_group}
\title{Evluating pESCA model when simulated parameters are avaliable}
\usage{
eval_metrics_simu_group(mu, A, B, S, ds, simulatedData)
}
\arguments{
\item{mu}{estimated offset term in column vector form}

\item{A}{estimated score matrix}

\item{B}{estimated loading matrix}

\item{S}{estimated group sparse pattern on \code{B}}

\item{ds}{a vector contains the number of variables in multiple data sets}

\item{simulatedData}{the output of function \code{dataSimu_group_sparse}}
}
\value{
This function returns a list contains \itemize{
\item  RVs_structs: a vector contains the RV coefficients in estimating
the global common (C123), local common (C12, C13, C23) and distinct structures
(D1, D2, D3);
\item  Ranks_structs: a vector contains the ranks of estimated
C123, C12, C13, C23, D1, D2, D3;
\item RMSEs_params: the relative mean squared errors (RMSEs) in estimating
the simulated parameters \eqn{\Theta, \Theta_1, \Theta_2, \Theta_3, \mu}
}
}
\description{
This function will evaluate the the performance of the constructed
pESCA model with group penalty when the simulated parameters are
avaliable.
}
\examples{
\dontrun{eval_metrics_simu_group(mu,A,B,S,ds,simulatedData)}

}
