% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FordWalford.R
\name{fordWalford}
\alias{fordWalford}
\title{Modelo de Ford-Walford.}
\usage{
fordWalford(
  dados,
  labelCX,
  labelCY,
  labelLX,
  labelLY,
  mainNameC,
  mainNameD,
  sexo = NULL
)
}
\arguments{
\item{dados}{Contém um dataframe com os dados importados.}

\item{labelCX}{Contém o rótulo para eixo x sobre média de ct.}

\item{labelCY}{Contém o rótulo para eixo y sobre média de ct.}

\item{labelLX}{Contém o rótulo para eixo x sobre o ln.}

\item{labelLY}{Contém o rótulo para eixo y sobre o ln.}

\item{mainNameC}{Contém o título do gráfico para regressão do ct.}

\item{mainNameD}{Contém o título do gráfico para regressão do ln.}

\item{sexo}{Contém um texto referente ao grupo atual.}
}
\description{
Esta função gera os primeiros parâmetros da história de vida dos peixes.
}
\examples{
fordWalford(dados, labelCX, labelCY, labelLX, labelLY, mainNameC, mainNameD, sexo)

}
