% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_functions.R
\name{gdp}
\alias{gdp}
\title{gdp penalty function}
\usage{
gdp(x, gamma, lambda)
}
\arguments{
\item{x}{a non-negative numeric value}

\item{gamma}{hyper-tuning parameter}

\item{lambda}{tuning parameter}
}
\value{
the value of gdp penalty
}
\description{
This function define the gdp penalty. The
formula is \code{gdp(x) =lambda*log(1+x/gamma)}.
Details can be found in \url{https://arxiv.org/abs/1807.04982}.
}
\examples{
\dontrun{gdp(0:9,gamma=1,lambda=5)}
}
