% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_functions.R
\name{gdp_sg}
\alias{gdp_sg}
\title{Super gradient of gdp penalty function}
\usage{
gdp_sg(x, gamma, lambda)
}
\arguments{
\item{x}{a non-negative numeric value}

\item{gamma}{hyper-tuning parameter}

\item{lambda}{tuning parameter}
}
\value{
the super-gradient of the gdp penalty function
}
\description{
This function define the super gradient of the gdp penalty.
The formula is \code{gdp_sg(x) = lambda/(gamma+x)}.
Details can be found in \url{https://arxiv.org/abs/1807.04982}
}
\examples{
\dontrun{gdp_sg(0:9,gamma=1,lambda=5)}
}
