% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{graficoR}
\alias{graficoR}
\title{Gráficos dos Modelos Matemáticos.}
\usage{
graficoR(ln_x, ln_y, regressao_ln, labelX, labelY, mainA, result, sexo = NULL)
}
\arguments{
\item{ln_x}{Contém o vetor com o eixo X do resultado de ln.}

\item{ln_y}{Contém o vetor com o eixo y do resultado de ln.}

\item{regressao_ln}{Contém o resultado da regressão de ln.}

\item{labelX}{Contém o rótulo para eixo x.}

\item{labelY}{Contém o rótulo para eixo y.}

\item{mainA}{Contém o título do gráfico.}

\item{result}{Contém a legenda do gráfico.}

\item{sexo}{Contém um texto referente ao grupo atual.}

\item{regressaoZ}{Contém o resultado da regressão da mortalidade (Z).}

\item{entradaX}{Contém o vetor com o eixo x.}

\item{entradaY}{Contém o vetor com o eixo y.}

\item{entradaB}{Contém o vetor com a linha da curva de Bertalanffy.}

\item{entradaG}{Contém o vetor com a linha da curva de Gompertz.}

\item{entradaL}{Contém o vetor com a linha da curva de Logistica.}

\item{emaB}{Contém o vetor com a linha da parte maior do IC para Bertalanffy.}

\item{emaG}{Contém o vetor com a linha da parte maior do IC para Gompertz.}

\item{emaL}{Contém o vetor com a linha da parte maior do IC para Logistica.}

\item{emeB}{Contém o vetor com a linha da parte menor do IC para Bertalanffy.}

\item{emeG}{Contém o vetor com a linha da parte menor do IC para Gompertz.}

\item{emeL}{Contém o vetor com a linha da parte menor do IC para Logistica.}

\item{Lin}{Contém as classes de comprimento e seu respectivo n agrupado.}

\item{LnNdt}{Contém vetor de conversão do numero total de peixes de cada classe de comprimento LN(N/dt).}

\item{taspas}{Contém a idade média de cada classe de comprimento (t').}

\item{agarraX}{Contém dados de x para calculo da curva de captura.}

\item{agarraY}{Contém dados de y para calculo da curva de captura.}

\item{labelZX}{Contém o rótulo para eixo x.}

\item{labelZY}{Contém o rótulo para eixo y.}

\item{modelo}{Contém o nome do modelo matemático.}

\item{mainE}{Contém o título do gráfico.}

\item{labX}{Contém o rótulo para eixo x.}

\item{labY}{Contém o rótulo para eixo y.}

\item{pausas}{Contém o valor para utilizar no break do histograma.}
}
\description{
Script que engloba todos os gráficos a serem apresentados posteriomente às análises.
}
\examples{
graficoR(ln_x, ln_y, regressao_ln, labelX, labelY, mainA, result, sexo=NULL)
graficoRLN(ln_x, ln_y, regressao_ln, labelX, labelY, mainA, result,sexo=NULL)
graficoN(entradaX, entradaY, entradaB, entradaG, entradaL, labelX, labelY, mainA)
graficoA(entradaX, entradaY, entradaB, entradaG, entradaL, emaB, emeB, emaG, emeG, emaL, emeL, labelX, labelY, mainA)
graficoZ(LiN, tAspas, lnNdT, agarraX, agarraY, labelZX, labelZY, mainA, modelo, sexo=NULL, mainE, labX, labY, pausas)

}
