% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Can not delete it}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{txt}
}
\value{
a string
}
\description{
I tried delete it, however the package can not pash check
}
\examples{
\dontrun{hello()}
}
