% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{index_Xi}
\alias{index_Xi}
\title{Index generating function}
\usage{
index_Xi(i, ds)
}
\arguments{
\item{i}{index for the ith data set.}

\item{ds}{a vector contains the number of variables for all
the data sets.}
}
\value{
A series of indexes for the variables in the ith data set
}
\description{
This function will generate a series of indexes used to index out
the variables of the ith data set when we only know the number of
variables
}
\examples{
\dontrun{index_Xi(2, c(400,200,100))}
}
