% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervaloConfianca.R
\name{intervaloConfianca}
\alias{intervaloConfianca}
\title{Intervalo de Confiança das Curvas Ajustadas.}
\usage{
intervaloConfianca(
  vbl,
  gom,
  lgc,
  curva_linha_ajustadaB,
  curva_linha_ajustadaG,
  curva_linha_ajustadaL,
  real_cont_fw
)
}
\arguments{
\item{vbl}{Contém dados completos do ajuste de Levenberg-Marquardt para parâmetros de Bertalanffy.}

\item{gom}{Contém dados completos do ajuste de Levenberg-Marquardt para parâmetros de Gompertz.}

\item{lgc}{Contém dados completos do ajuste de Levenberg-Marquardt para parâmetros de Logística.}

\item{curva_linha_ajustadaB}{Contém dados da curva de Bartalanffy com ajuste de Levenberg-Marquardt.}

\item{curva_linha_ajustadaG}{Contém dados da curva de Gompertz com ajuste de Levenberg-Marquardt.}

\item{curva_linha_ajustadaL}{Contém dados da curva de Logística com ajuste de Levenberg-Marquardt.}

\item{real_cont_fw}{Contém o "n" da amostra.}
}
\description{
Script que calcula o intervalo de confiança sobre os dados ajustados pelo modelo de Levenberg-Marquardt.
}
\examples{
intervaloConfianca(vbl, gom, lgc, curva_linha_ajustadaB, curva_linha_ajustadaG, curva_linha_ajustadaL, real_cont_fw)

}
