% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{inver_logit}
\alias{inver_logit}
\title{Element-wise inverse logit function}
\usage{
inver_logit(x)
}
\arguments{
\item{x}{a matrix contains real numbers}
}
\value{
a probability matrix
}
\description{
This function will do inveser logit transformation on a matrix
contains real numbers. The formula is \code{inver_logit(x) = (1+exp(-x))^(-1)}
}
\examples{
\dontrun{inver_logit(matrix(rnorm(3*4),nrow=3,ncol=4))}

}
