% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{log10_seq}
\alias{log10_seq}
\title{Generate log-spaced sequence}
\usage{
log10_seq(from, to, length.out)
}
\arguments{
\item{from}{The initial value of the sequence}

\item{to}{The last value of the sequence}

\item{length.out}{desired length of the sequence}
}
\value{
A \code{length.out} dimensional squence
}
\description{
This function will generate a log-spaced sequence.
The inputs are the same as function \code{seq} in base library.
}
\examples{
\dontrun{log10_seq(from=1, to=500, length.out=30)}

}
