% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_partitions_functions.R
\name{log_part_B_g}
\alias{log_part_B_g}
\title{The first order gradient of log-partation function of Bernoulli distribution}
\usage{
log_part_B_g(Theta)
}
\arguments{
\item{Theta}{a matrix of natural parameter}
}
\value{
value of the first order gradient
}
\description{
This function define first order gradient of the log-partation function
of Bernoulli distribution. The formula is
\code{log_part_B_g(Theta) = exp(Theta)/(1+exp(Theta))}.
Details can be found in \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{log_part_B_g(matrix(0:9,2,5))}
}
