% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{logit}
\alias{logit}
\title{Element-wise logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{a matrix contains probabilities}
}
\value{
The logit transformation of a probability matrix
}
\description{
This function will do logit transformation on a probability matrix.
The formula is \code{logit(x) = log(x/(1-x))}
}
\examples{
\dontrun{logit(matrix(rbeta(3*4,1,1),nrow=3,ncol=4))}

}
