% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_functions.R
\name{lq}
\alias{lq}
\title{Lq penalty function}
\usage{
lq(x, gamma, lambda)
}
\arguments{
\item{x}{a non-negative numeric value}

\item{gamma}{hyper-tuning parameter}

\item{lambda}{tuning parameter}
}
\value{
the value of the Lq penalty
}
\description{
This function define the Lq penalty. The
formula is \code{Lq(x) =lambda*x^q}.
Details can be found in \url{https://arxiv.org/abs/1807.04982}
}
\examples{
\dontrun{lq(0:9,gamma=0.5,lambda=5)}
}
