% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_functions.R
\name{lq_sg}
\alias{lq_sg}
\title{Super gradient of the Lq penalty function}
\usage{
lq_sg(x, gamma, lambda)
}
\arguments{
\item{x}{a non-negative numeric value}

\item{gamma}{hyper-tuning parameter}

\item{lambda}{tuning parameter}
}
\value{
the super-gradient of the Lq penalty function
}
\description{
This function define the super gradient of the Lq penalty.
The formula is \code{lq_sg(x) = lambda * q * x^(p-1)}.
Details can be found in \url{https://arxiv.org/abs/1807.04982}
}
\examples{
\dontrun{
lq_sg(0:9,gamma=1,lambda=5)
lq_sg(0:9,gamma=0.5,lambda=5)
}
}
