% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalidadeZ.R
\name{mortalidadeZ}
\alias{mortalidadeZ}
\title{Mortalidade Total (Z).}
\usage{
mortalidadeZ(
  c_infinito,
  k,
  tzero,
  dados,
  real_cont_fw,
  idioma,
  modelo,
  mainE,
  labX,
  labY,
  sexo = NULL,
  label
)
}
\arguments{
\item{c_infinito}{Contém o comprimento assintótico.}

\item{k}{Contém a taxa de crescimento.}

\item{tzero}{Contém o tempo no comprimento zero.}

\item{dados}{Contém o dataframe com os dados importados.}

\item{real_cont_fw}{Contém do "n" da amostra.}

\item{idioma}{Contém o idioma do resultado.}

\item{modelo}{Contém o modelo usado para cálculos.}

\item{mainE}{Contém o título do gráfico.}

\item{labX}{Contém o rótulo para eixo x.}

\item{labY}{Contém o rótulo para eixo y.}

\item{sexo}{Contém um texto referente ao grupo atual.}

\item{label}{Contém a legenda referente ao grupo atual.}
}
\description{
Script que calcula a mortalidade total partindo de parâmetros de Ford-Walford ou de Levenberg-Marquardt.
}
\examples{
mortalidadeZ(c_infinito, k, tzero, dados, real_cont_fw, idioma, modelo, mainE, labX, labY, sexo, label)

}
