% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{obj_logistic}
\alias{obj_logistic}
\title{Logistic loss function}
\usage{
obj_logistic(X, Theta)
}
\arguments{
\item{X}{a binary matrix}

\item{Theta}{a natual parameter (log-odds) matrix}
}
\value{
The logistic loss
}
\description{
This function will compute the logistic loss when a
binary data set and its natural parameter matrix is avaliable.
}
\examples{
\dontrun{
Theta <- matrix(rnorm(3*4),3,4)
X <- matrix(data=0,3,4)
X[Theta>0] <- 1
obj_logistic(X,Theta)
}
}
