% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{ones}
\alias{ones}
\title{ones function}
\usage{
ones(n)
}
\arguments{
\item{n}{a integer number indicates the dimension of the vector}
}
\value{
a n dimensional column vector contains ones
}
\description{
This function will generate a column of ones
}
\examples{
\dontrun{ones(10)}

}
