% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_algorithms.R
\name{pESCA}
\alias{pESCA}
\title{Penalized exponential family simultaneous component analysis (pESCA) model}
\usage{
pESCA(dataSets, dataTypes, lambdas, penalty = "L2",
  fun_concave = "gdp", opts = list())
}
\arguments{
\item{dataSets}{a list contains multiple matrices with same number of rows.
Each matrix (\code{samples * variables}) indicates a data set.}

\item{dataTypes}{a string indicates the data types of the multiple data sets.}

\item{lambdas}{a numeric vector indicates the values of tuning parameters for
each data set.}

\item{penalty}{The name of the penalty you want to used. \itemize{
\item "L2": group-wise concave L2 norm penalty;
\item "L1": group-wise concave L1 norm penalty;
\item "element": element-wise concave penalty;
\item "composite": the composition of group- and element-wise penalty.
}}

\item{fun_concave}{a string indicates the used concave function. Three options
are included in the algorithm. \itemize{
\item "gdp": GDP penalty;
\item "lq": Lq penalty;
\item "scad": SCAD penalty.
}}

\item{opts}{a list contains the options of the algorithms. \itemize{
\item tol_obj: tolerance for relative change of joint loss function, default:1E-6;
\item maxit: max number of iterations, default: 1000;
\item gamma: hyper-parameter of the concave penalty, default: 1;
\item R: the initial number of PCs, default: 0.5 \code{0.5*min(I,J)};
\item rand_start: initilization method, random (1), SCA(0), default: 0;
\item alphas: dispersion parameters of exponential dispersion families, default: 1.
\item thr_path: the option to generate thresholding path, default: 0;
\item quiet: quiet==1, not show the progress when running the algorithm, default: 0.
}}
}
\value{
This function returns a list contains the results of a pESCA mdoel. \itemize{
\item mu: offset term;
\item A: score matrix;
\item B: loading matrix;
\item S: group sparse pattern of B;
\item varExpTotals: total variation explained of each data set and the full data set;
\item varExpPCs: variation explained of each data set and each component;
\item Sigmas: the group length (the definition depends on the used input type). Only meaningful
for group-wise sparisty;
\item iter: number of iterations;
\item diagnose$hist_objs: the value of loss function pluse penalty at each iteration;
\item diagnose$f_objs: the value of loss function at each iteration;
\item diagnose$g_objs: the value of penalty function at each iteration;
\item diagnose$rel_objs: relative change of diagnose$hist_obj at each iteration;
\item diagnose$rel_Thetas: relative change of Theta at each iteration.
}
}
\description{
This is the main function for construncting a pESCA model on multiple data
sets. The potential different data types in these data sets are tackled by
the assumption of exponential family distribution. Gaussian for quantitative
data, Bernoulli for binary data and Poisson for count data. Although the option
for count data using Poisson distribution is included in the algorithm, we recommend
to do variance stabilizing transformation on the count data, such as Next-Gen
sequencing data, and then use the transformed data as quantitative data sets. The
details of the developed algorithm can be found in \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{
# Suppose we have three data sets X1, X2, X3
# They are quantitative, quantitative and binary matrices
pESCA(dataSets = list(X1, X2, X3),
              dataTypes = 'GGB',
              lambdas = c(20, 20, 10),
              penalty = 'L2',
              fun_concave = 'gdp',
              opts = list())
}

}
