% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_CV_twoSteps.R
\name{pESCA_CV_twoSteps}
\alias{pESCA_CV_twoSteps}
\title{pESCA model selection based on cross validation error}
\usage{
pESCA_CV_twoSteps(dataSets, dataTypes, lambdas_g, lambdas_b,
  penalty = "L2", fun_concave = "gdp", opts = list())
}
\arguments{
\item{dataSets}{a list contains multiple matrices with same number of rows.
Each matrix (\code{samples * variables}) indicates a data set.}

\item{dataTypes}{a string indicates the data types of the multiple data sets.}

\item{lambdas_g}{a vector cotains a sequence of values of lambda for
the loadings related to quantitative data sets}

\item{lambdas_b}{a vector cotains a sequence of values of lambda for
the loadings related to binary data sets}

\item{penalty}{The name of the penalty you want to used. \itemize{
\item "L2": group-wise concave L2 norm penalty;
\item "L1": group-wise concave L1 norm penalty;
\item "element": element-wise concave penalty;
\item "composite": the composition of group- and element-wise penalty.
}}

\item{fun_concave}{a string indicates the used concave function. Three options
are included in the algorithm. \itemize{
\item "gdp": GDP penalty;
\item "lq": Lq penalty;
\item "scad": SCAD penalty.
}}

\item{opts}{a list contains the options of the algorithms. \itemize{
\item tol_obj: tolerance for relative change of joint loss function, default:1E-6;
\item maxit: max number of iterations, default: 1000;
\item gamma: hyper-parameter of the concave penalty, default: 1;
\item R: the initial number of PCs, default: 0.5 \code{0.5*min(I,J)};
\item rand_start: initilization method, random (1), SCA(0), default: 0;
\item alphas: dispersion parameters of exponential dispersion families, default: 1.
\item thr_path: the option to generate thresholding path, default: 0;
\item quiet: quiet==1, not show the progress when running the algorithm, default: 0.
}}
}
\value{
This function returns a list contains the results of a pESCA mdoel. \itemize{
\item opts_opt: selected lambdas;
\item opts_opt: the initialization of the selected model;
\item cvErrors_b: the CV errors during the tuning of lambda for
loadings related to binary data sets;
\item cvErrors_g: the CV errors during the tuning of lambda for
loadings related to quantitative data sets;
}
}
\description{
This function implements a missing value based CV model selection
approach for the pESCA model on mutliple data sets with different
data types, such as quantitative and binary.
The details can be found in  \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{
result_CV <- pESCA_CV_twoSteps(dataSets, dataTypes,
                               lambdas_g, lambdas_b,
                               penalty='L2', fun_concave='gdp', opts=opts)
}

}
