% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedorObjetos.R
\name{removedor}
\alias{removedor}
\title{Removedor de Objetos/variáveis do ambiente.}
\usage{
removedor(variaveis)
}
\arguments{
\item{variaveis}{Indica um número referente ao conjunto de variáveis que será removida.}
}
\description{
Script que remove as variáveis para que não ocorra conflitos no uso em dados agrupados.
}
\examples{
removedor(variaveis)

}
