% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remoção de Outliers.}
\usage{
removeOutliers(Dados)
}
\arguments{
\item{dados}{Contém um dataframe com os dados importados.}
}
\description{
Script que analisa e remove estatisticamente outliers.
}
\examples{
removeOutliers(dados)

}
