% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPescaMain.R
\name{rPesca}
\alias{rPesca}
\title{Interface rPesca.}
\usage{
rPesca(
  cores = "1",
  idioma,
  un,
  tipoComprimento = "Total",
  tempo,
  tipo_dados,
  nome_dados,
  planilha = NULL,
  separador = NULL,
  grupo = NULL,
  nomeA = NULL,
  nomeB = NULL,
  Fonte = NULL
)
}
\arguments{
\item{cores}{Determina se será exibido o plot 1 = em escala de cinza ou 2 = colorido. (Por padrão: Escala de cinza)}

\item{idioma}{Determina o idioma a ser apresentado no resultado, sendo 1 = português, 2 = inglês.(Por padrão: Português)}

\item{un}{Determina a unidade de medida 1 = cm, 2 = mm, 3 = m.(Por padrão: cm)}

\item{tipoComprimento}{Determina o tipo do comprimento aplicado, exemplo: se informar "total" será apresentado como comprimento total.}

\item{tempo}{Determina a medida da idade 1 = ano, 2 = mês, 3 = dia.(Por padrão: Ano)}

\item{tipo_dados}{Informa a extensão do arquivo com os dados, sendo 1 = xlsx, 2 = csv, 3 = txt, 4 = dataframe.(Por padrão: Dataframe)}

\item{nome_dados}{Campo character para informar nome do arquivo que está com os dados.(Preenchimento obrigatório)}

\item{planilha}{Campo character para informar a pasta com os dados dentro da planilha xlsx.(Por padrão: Primeira pasta)}

\item{separador}{Campo character para informar o separador utilizado caso o arquivo esteja em csv.(Por padrão: Vírgula)}

\item{grupo}{Informa se deseja analisar por agrupamento, sendo 1 = por agrupamento e o não preenchimento = sem agrupamento.}

\item{nomeA}{Informa o nome do grupo A.}

\item{nomeB}{Informa o nome do grupo B.}

\item{Fonte}{Informa a família da fonte a ser utilizada nos gráficos.}
}
\description{
Script que integra todos os demais. Inicia qualquer análise de acordo com modelos de crescimento, mortalidade e verossimilhança.
}
\examples{
rPesca(cores, idioma, un, tipoComprimento, tempo, tipo_dados, nome_dados, planilha, separador, grupo, nomeA, nomeB, Fonte)
rPesca(2, 1, 1, "Total", 1, 1, "dados.xlsx", "ara")
rPesca(2, 2, 1, "Total", 1, 1, "pomsal-new.xlsx", "TESTE 5", NULL, 1, "m", "f")

}
