% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{RV_modified}
\alias{RV_modified}
\title{Modified RV coefficient of two matrices}
\usage{
RV_modified(X, Y)
}
\arguments{
\item{X}{a matrix}

\item{Y}{another matrix}
}
\value{
This function returns the modified RV coefficient between
two matrices
}
\description{
This function will compute the modified RV coefficient of two
matrices. The details of the modified RV coefficient can be found
in the paper \url{https://academic.oup.com/bioinformatics/article/25/3/401/244239}.
}
\examples{
\dontrun{RV_modified(X,Y)}
}
