% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_functions.R
\name{scad_sg}
\alias{scad_sg}
\title{Super gradient for the SCAD penalty function}
\usage{
scad_sg(x, gamma, lambda)
}
\arguments{
\item{x}{a non-negative numeric value}

\item{gamma}{hyper-tuning parameter}

\item{lambda}{tuning parameter}
}
\value{
the super gradient of the SCAD penalty
}
\description{
This function define the supergradient of the SCAD penalty.
Details can be found in \url{https://arxiv.org/abs/1807.04982}
}
\examples{
\dontrun{scad_sg(0:9,gamma=3.7,lambda=3)}
}
