% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_estimation_procedures.R
\name{svd_CV}
\alias{svd_CV}
\title{Model selection of a svd model using missing value based CV error}
\usage{
svd_CV(X, K = 3, Rs = 1:15, opts = list(), ratio_mis = 0.1)
}
\arguments{
\item{X}{a quantitative data set}

\item{K}{K-fold cross validation}

\item{Rs}{the searching range of the number of components}

\item{opts}{a list contains the setting for the algorithm. \itemize{
\item tol_obj: tolerance for relative change of hist_obj, default:1E-6;
\item maxit: max number of iterations, default: 1000;
}}

\item{ratio_mis}{the propotion of missing values}
}
\value{
This function returns a list contains \itemize{
\item Rs: the number of PCs used for model selection
\item cvErrors: K-fold cross validation errors
}
}
\description{
This function implemented a missing value based CV model selection approach.
First, ratio_mis percent elements are randomly selected as missing values. After
that a EM-SVD model is constructed to estimate the prediction error.
The details of this function can be found in \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{svd_CV(X,K=3,Rs = 1:15,opts=list(),ratio_mis=0.1)}

}
