% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_simu_element_sparse.R
\name{test_dataSimu_element}
\alias{test_dataSimu_element}
\title{Test data simulation only with element-wise sparse pattern}
\usage{
test_dataSimu_element(n, ds, R, dataTypes)
}
\arguments{
\item{n}{the number of objects}

\item{ds}{a vector for the number of variables in each data set}

\item{R}{number of simulated PCs}

\item{dataTypes}{a string indicates the data type of each data set,
possible options include 'G': Gaussian, 'B': Bernoulli.}
}
\value{
This function returns a list contains all the test results
}
\description{
This function will test if the simulated data sets have the proper
properties, such as correct data types, orghogonal score matrix,
score matrix has 0 column means, element-wise sparsity, with correct SNRs.
}
\examples{
\dontrun{
test_results_GGG <- test_dataSimu_element(n=n, ds=ds,R=3
                                dataTypes = 'GGG')
test_results_BBB <- test_dataSimu_element(n=n, ds=ds,R=3
                                dataTypes = 'BBB')
}
}
