% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESCA_simu_group_sparse.R
\name{test_dataSimu_group}
\alias{test_dataSimu_group}
\title{Test data simulation with group sparse pattern}
\usage{
test_dataSimu_group(n, ds, dataTypes)
}
\arguments{
\item{n}{the number of objects}

\item{ds}{a vector for the number of variables in each data set}

\item{dataTypes}{a string indicates the data type of each data set,
possible options include 'G': Gaussian, 'B': Bernoulli.}
}
\value{
This function returns a list contains all the test results
}
\description{
This function will test if the simulated data sets have the proper
properties, such as correct data types, orghogonal score matrix,
score matrix has 0 column means, group sparsity, with correct SNRs.
}
\examples{
\dontrun{
test_results_GGG <- test_dataSimu_group(n=n, ds=ds,
                                dataTypes = 'GGG')
test_results_BBB <- test_dataSimu_group(n=n, ds=ds,
                                dataTypes = 'BBB')
}
}
