% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{trace_fast}
\alias{trace_fast}
\title{A function to compute the trace of two matrices product}
\usage{
trace_fast(X, Y)
}
\arguments{
\item{X}{a numerical matrix}

\item{Y}{a numerical matrix with the same size as \code{X}}
}
\value{
This function returns a scalar contains the trace
}
\description{
This function will compute the trace of two matrices
}
\examples{
\dontrun{trace_fast(X, Y)}
}
