% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_pESCA_algorithm.R
\name{update_B_L1}
\alias{update_B_L1}
\title{Updating loading matrix B with conave L1 norm penalty}
\usage{
update_B_L1(JHk, A, B0, Sigmas0, d, fun_concave, alphas, rhos, lambdas,
  gamma)
}
\arguments{
\item{JHk}{An output of the majorizaiotn step}

\item{A}{The score matrix A during k-th iteration}

\item{B0}{The loading matrix B during the previous iteration}

\item{Sigmas0}{The group length during the previous iteration}

\item{d}{A numeric vector contains the numbers of variables in different data sets}

\item{fun_concave}{A string indicates the used concave function}

\item{alphas}{The dispersion parameters of exponential dispersion families}

\item{rhos}{An output of the majorizaiotn step}

\item{lambdas}{A numeric vector indicates the values of tuning parameters for
each data set.}

\item{gamma}{The hyper-parameter of the concave penalty}
}
\value{
This function returns the updated loading matrix B.
}
\description{
This is an intermediate step of the algorithm for fitting pESCA model. The
details of this function can be found in ref thesis.
}
\examples{
\dontrun{
B <- update_B_L1(JHk,A,B0,Sigmas0,d,
                   fun_concave,alphas,rhos,lambdas,gamma)
}
}
