% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{varExp_Gaussian}
\alias{varExp_Gaussian}
\title{Compute the variation expalined raitios when Gaussian data is used}
\usage{
varExp_Gaussian(X, mu, A, B, Q)
}
\arguments{
\item{X}{a \eqn{n*d} quantitative data set}

\item{mu}{a \eqn{n*1} column offset term}

\item{A}{a \eqn{n*R} score matrix}

\item{B}{a \eqn{d*R} loading matrix}

\item{Q}{a \eqn{n*d} weighting matrix of the same size of X}
}
\value{
This function returns a list contains the variaiton expalined
ratios of the whole model (varExp_total) or of each component (varExp_PCs).
}
\description{
This function computes the variation expalined ratios for component
model on quantitative data sets. Details can be found
in the paper \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{ out <- varExp_Gaussian(X,mu,A,B,Q)
          out$varExp_total
          out$varExp_PCs
}
}
