% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPescaMain.R
\name{rPesca}
\alias{rPesca}
\title{Interface rPesca.}
\usage{
rPesca(
  cores = 1,
  idioma = 1,
  un = 1,
  tipoComprimento = "Total",
  tempo = 1,
  tipo_dados = 1,
  nome_dados,
  planilha = NULL,
  separador = NULL,
  grupo = NULL,
  nomeA = NULL,
  nomeB = NULL,
  Fonte = NULL
)
}
\arguments{
\item{cores}{Determina como será exibido o gráfico, sendo 1 = em escala de cinza ou 2 = colorido. (Por padrão: Escala de cinza)}

\item{idioma}{Determina o idioma a ser apresentado no resultado, sendo 1 = português ou 2 = inglês.(Por padrão: Português)}

\item{un}{Determina a unidade de medida, sendo 1 = cm, 2 = mm ou 3 = m.(Por padrão: cm)}

\item{tipoComprimento}{Determina o tipo do comprimento aplicado, exemplo: se informar "Total" será apresentado como Comprimento Total.}

\item{tempo}{Determina a medida da idade, sendo 1 = ano, 2 = mês ou 3 = dia.(Por padrão: Ano)}

\item{tipo_dados}{Informa a extensão do arquivo com os dados, sendo 1 = xls/xlsx, 2 = csv, 3 = txt (primeira linha é cabeçalho) ou 4 = dataframe.(Por padrão: Dataframe)}

\item{nome_dados}{Campo caracter para informar nome do arquivo que está com os dados incluindo sua extensão. Exemplo: "dados.xls". (Preenchimento obrigatório)}

\item{planilha}{Campo caracter para informar a pasta com os dados dentro da planilha xls/xlsx.(Por padrão: Primeira pasta)}

\item{separador}{Campo caracter para informar o separador utilizado caso o arquivo esteja em csv. Exemplo: para espaço " ", para vírgula ",", etc. (Por padrão: Vírgula)}

\item{grupo}{Informa se deseja analisar em grupos, sendo 1 = sim ou sem preenchimento = não.}

\item{nomeA}{Informa o nome do grupo A.(Obrigatório para análise de grupos)}

\item{nomeB}{Informa o nome do grupo B.(Obrigatório para análise de grupos)}

\item{Fonte}{Informa a família da fonte a ser utilizada nos gráficos, sendo NULL = Times New Roman, 1 = Arial ou 2 = Courier New.(Por padrão: Times New Roman)}
}
\description{
Script que análisa dados amostrais de acordo com modelos de crescimento, mortalidade e verossimilhança.
}
\examples{
data("exemplo")
rPesca(2, 1, 1, "Total", 1, 4, ex_individual)

data("exemplo")
rPesca(2, 1, 1, "Total", 1, 4, ex_grupo, NULL, NULL, 1, "ara", "ita")

data("example")
rPesca(2, 2, 1, "Total", 1, 4, eg_individual)

data("example")
rPesca(2, 2, 1, "Total", 1, 4, eg_group, NULL, NULL, 1, ara, ita)

}
